ChangedSCInfosFromDateTimeToDouble

+++ Backend/Dtos/TlinquiryDto.cs
@@ -8,9 +8,9 @@ public class TlinquiryDto
     [Required] public string InvoiceOn { get; set; } = null!;
     [Required] public string RetrieveDate { get; set; } = null!;
     [Required] public string RetrieveLocation { get; set; } = null!;
-    [Required] public int SCGeneral { get; set; }
-    [Required] public int SCMain { get; set; }
-    [Required] public int SCTrail { get; set; }
+    [Required] public double SCGeneral { get; set; }
+    [Required] public double SCMain { get; set; }
+    [Required] public double SCTrail { get; set; }
     [Required] public string PortOfDeparture { get; set; } = null!;
     [Required] public string Ets { get; set; } = null!;
     [Required] public string Eta { get; set; } = null!;
